Removed upcoming events & line chart
