--- Admin-app/src/components/Charts/LineChart.test.js
@@ -1,11 +1,0 @@
-import { render, screen, cleanup } from "@testing-library/react";
-import "@testing-library/jest-dom";
-import LineChart from "./LineChart";
-
-test("LineChart should render without errors", () => {
-  render(<LineChart />);
-});
-
-afterEach(() => {
-  cleanup();
-});
