--- Admin-app/src/components/UpcomingEvents/UpcomingEvents.css
@@ -1,75 +1,0 @@
-.UPCOMING-EVENTS__div {
-    margin-top: 4%;
-}
-
-.UPCOMING-EVENTS-1__background {
-    background-color: var(--background-main);
-    border-radius: 20px;
-}
-
-.UPCOMING-EVENTS__date {
-    background: var(--gradient-to-right);
-    color: var(--text-primary);
-    width: 90px;
-    height: 90px;
-    text-align: center;
-    border-radius: 15px;
-    position: relative;
-    bottom: 20px;
-    padding-top: 1.5rem;
-    margin-left: 20px;
-}
-
-.UPCOMING-EVENTS__day {
-    font-size: 30px;
-    font-weight: bold;
-    padding-top: 0.5rem;
-}
-
-.UPCOMING-EVENTS__month {
-    font-size: 20px;
-    text-transform: uppercase;
-}
-
-.UPCOMING-EVENTS__title {
-    color: var(--text-primary);
-    font-weight: bold;
-    margin-left: 130px;
-    font-size: 20px;
-    margin-top: -70px;
-}
-
-.UPCOMING-EVENTS__time {
-    color: var(--text-inactive);
-    font-weight: bold;
-    margin-left: 130px;
-    font-size:  15px;
-    margin-top: 10px;
-}
-
-.UPCOMING-EVENTS__viewMore {
-    color: var(--text-inactive) !important;
-    font-size: 15px !important;
-    font-weight: bold !important;
-    margin-left: 123px !important;
-}
-
-.UPCOMING-EVENTS__arrow {
-    position: relative;
-    left: 5px;
-    bottom: 1px;
-}
-
-.UPCOMING-EVENTS__modal__header {
-    font-weight: bold;
-    font-size: 20px;
-}
-
-.UPCOMING-EVENTS__modal__desc {
-    font-size: 15px;
-}
-
-.UPCOMING-EVENTS__viewMore:hover {
-    background-color: transparent !important;
-    color: var(--primary-main) !important;
-}
