+++ Admin-app/src/screens/Dashboard/Dashboard.css
@@ -13,16 +13,12 @@ a {
   color: var(--text-primary);
 }
 
-.PATIENT-LIST__title{
-  font-family: "Roboto", "Segoe UI", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
+.PATIENT-LIST__title {
+  font-family: "Roboto", "Segoe UI", "Oxygen", "Ubuntu", "Cantarell",
+    "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
   text-align: center !important;
   font-size: 16px !important;
   font-weight: bold !important;
   border-bottom: 1px solid black;
   margin-top: 10px !important;
-}
-
-.PATIENT-LIST {
-  border-radius: 20px;
-  width: 30% !important;
 }
